Decompile func0f0bc478

+++ src/include/game/game_0b63b0.h
@@ -49,7 +49,7 @@ void currentPlayerSurroundWithExplosions(s32 arg0);
 u32 func0f0bc120(void);
 void func0f0bc3f0(void);
 s16 func0f0bc44c(void);
-u32 func0f0bc478(void);
+s16 func0f0bc478(void);
 u32 func0f0bc4c0(void);
 u32 func0f0bc554(void);
 u32 func0f0bc7ac(void);
